split vis builder and vis gl into sep files
