+++ src/main.rs
@@ -5,9 +5,10 @@ mod globe;
 mod icosphere;
 mod mouse;
 mod points;
-mod vis;
+mod vis_build;
 mod vis_ctx;
-use vis::VisBuilder;
+mod vis_gl;
+use vis_build::VisBuilder;
 
 pub trait VisState {
     fn update_points(&mut self, ms: f32) -> Vec<f32>;
